refactor: move all evaluations submitted check to helper function
